// app/components/UI/Predict/types/polymarket.ts
export type OrderResponse = {
  errorMsg: string;
  makingAmount: string;
  orderID: string;
  status: string;
  success: boolean;
  takingAmount: string;
  transactionsHashes: string[];
};